// repos/epaper-stock-price/src/models/interfaces.ts
export interface ISingleConfig {
  stream: string;
  market: string;
  symbol: string;
  interval: string;
  symbolType: number;
  currency: string;
  ledBrightness: number;
  alertEnabled?: boolean;
  multiplierEnabled?: boolean;
  multiplier?: number;
  alertConfig?: {
    alertEnabled?: boolean;
    triggerType: string;
    triggerValue: number;
    flashLightbar: boolean;
    playSound: boolean;
    soundType: string;
    changeLightBarColor: boolean;
    lightBarColor: string;
  };
  gainTrackingEnabled: boolean;
  purchasePrice?: number;
  noOfStocks: number;
  showFullAssetValue: boolean;
  isShortSell: boolean;
  extraConfig: {
    aggregateTime: string;
    isMetalCommodity: boolean;
    unit: string;
    divideBy100?: boolean;
    manualSearch: boolean;
  };
}

export interface IPlaylistConfig {
  name: string;
  cycleInterval: string;
  updateInterval: string;
  ledBrightness: number;
  cycleMode: string;
  isCalculateOnDaily: boolean;
  symbols: {
    stream?: string;
    market?: string;
    symbol: string;
    symbolType: number;
    currency: string;
    gainTrackingEnabled: boolean;
    purchasePrice?: number;
    noOfStocks: number;
    showFullAssetValue: boolean;
    isShortSell: boolean;
    multiplierEnabled?: boolean;
    multiplier?: number;
    extraConfig: {
      aggregateTime: string;
      isMetalCommodity: boolean;
      unit: string;
      divideBy100: boolean;
      manualSearch: boolean;
      feature?: string;
      humbl3Id?: string;
      musicCreatorURL?: string;
      displaySource?: string;
      displayLabel?: string;
      displayTimeFrame?: string;
      title?: string;
      subTitle?: string;
    };
  }[];
}

export interface IDevice {
  _id?: string;
  name?: string;
  macAddress: string;
  isActive: boolean;
  userId?: string;
  registration?: {
    key: string;
    valid: Date;
  };
  mode: string; //text, page
  isPlaylist?: boolean;
  pageConfig?: {
    pageId: number;
    ledColor: string;
  }[];
  config?: ISingleConfig | {};
  playlist?: IPlaylistConfig;
  playlistId?: string;
  firmwareVersion?: string;
  lastStatusOn: DateTime | string;
  batteryStatus: number;
  nightMode?: boolean;
  nightModeStart?: string;
  nightModeEnd?: string;
  extras?: any;
  createdAt?: DateTime;
  updatedAt?: DateTime;
}

export interface IDeviceWithUser extends Omit<IDevice, "userId"> {
  user: {
    _id: string;
    name: string;
    email: string;
    timeZone: string;
    timeFormat: string;
  };
}